refactor panToTarget to work with annotations and content search
